Varios

+++ Models/CondicionRegla.cs
@@ -4,7 +4,7 @@ namespace AlarmaDisparadorCore.Models
     {
         public int Id { get; set; }
         public int IdRegla { get; set; }
-        public int IdValor { get; set; }
+        public short IdValor { get; set; }
         public string Operador { get; set; }
         public string Valor { get; set; }
     }
